+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * **trader:** added estimation of taker fees for better PnL calculation.
 * **trader:** limited decimal places for numbers in the web UI.
+* **trader:** added upgrade path for saved data.
 
 ## [1.1.0] (2021-08-09)